Create app info section
Reuse and remove "switch safety" component.
Align CSS in shared offer page components.
