+++ src/client/pages/OfferNew/index.tsx
@@ -4,12 +4,7 @@ import React from 'react'
 import { Redirect, useHistory, useRouteMatch } from 'react-router'
 import { LoadingPage } from 'components/LoadingPage'
 import { TopBar } from 'components/TopBar'
-import {
-  getIsoLocale,
-  useCurrentLocale,
-  Market,
-  useMarket,
-} from 'components/utils/CurrentLocale'
+import { getIsoLocale, useCurrentLocale } from 'components/utils/CurrentLocale'
 import { Page } from 'components/utils/Page'
 import { SessionTokenGuard } from 'containers/SessionTokenGuard'
 import {
@@ -17,14 +12,14 @@ import {
   useQuoteBundleQuery,
   useRedeemedCampaignsQuery,
 } from 'data/graphql'
-import { SwitchSafetySection } from 'pages/OfferNew/SwitchSafetySection'
-import { getOfferData } from 'pages/OfferNew/utils'
 import { useVariation, Variation } from 'utils/hooks/useVariation'
 import { trackOfferGTM } from 'utils/tracking/gtm'
 import { getUtmParamsFromCookie, TrackAction } from 'utils/tracking/tracking'
 import { localePathPattern } from 'l10n/localePathPattern'
 import { useQuoteIds } from '../../utils/hooks/useQuoteIds'
 import { LanguagePicker } from '../Embark/LanguagePicker'
+import { getOfferData } from './utils'
+import { AppInfoSection } from './AppInfoSection'
 import { Checkout } from './Checkout'
 import { FaqSection } from './FaqSection'
 import { Introduction } from './Introduction'
@@ -43,7 +38,6 @@ const createToggleCheckout = (history: History<any>, locale?: string) => (
 export const OfferNew: React.FC = () => {
   const currentLocale = useCurrentLocale()
   const localeIsoCode = getIsoLocale(currentLocale)
-  const currentMarket = useMarket()
   const variation = useVariation()
   const history = useHistory()
   const { data: redeemedCampaignsData } = useRedeemedCampaignsQuery()
@@ -125,7 +119,7 @@ export const OfferNew: React.FC = () => {
               )}
             </TrackAction>
             <Perils offerData={offerData} />
-            {currentMarket !== Market.Dk && <SwitchSafetySection />}
+            <AppInfoSection />
             <FaqSection />
             <Checkout
               offerData={offerData}
